[FLINK-23704][task] Extract LatencyMarkerEmitter class
